Auslagern der Scenen in eigen Klassen für mehr übersichtlichkeit
